Refactor send for review notification
