+++ src/main/java/uk/gov/hmcts/divorce/citizen/event/CitizenInviteApplicant2.java
@@ -7,11 +7,11 @@ import uk.gov.hmcts.ccd.sdk.api.CCDConfig;
 import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
 import uk.gov.hmcts.ccd.sdk.api.ConfigBuilder;
 import uk.gov.hmcts.ccd.sdk.api.callback.AboutToStartOrSubmitResponse;
-import uk.gov.hmcts.divorce.citizen.notification.ApplicationSentForReviewApplicant1Notification;
-import uk.gov.hmcts.divorce.citizen.notification.ApplicationSentForReviewApplicant2Notification;
+import uk.gov.hmcts.divorce.citizen.notification.ApplicationSentForReviewNotification;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
 import uk.gov.hmcts.divorce.divorcecase.model.State;
 import uk.gov.hmcts.divorce.divorcecase.model.UserRole;
+import uk.gov.hmcts.divorce.notification.NotificationDispatcher;
 
 import java.time.LocalDate;
 import java.time.temporal.ChronoUnit;
@@ -31,10 +31,10 @@ public class CitizenInviteApplicant2 implements CCDConfig<CaseData, State, UserR
     public static final String CITIZEN_INVITE_APPLICANT_2 = "citizen-invite-applicant2";
 
     @Autowired
-    private ApplicationSentForReviewApplicant1Notification applicationSentForReviewApplicant1Notification;
+    private ApplicationSentForReviewNotification applicationSentForReviewNotification;
 
     @Autowired
-    private ApplicationSentForReviewApplicant2Notification applicationSentForReviewApplicant2Notification;
+    private NotificationDispatcher notificationDispatcher;
 
     @Override
     public void configure(final ConfigBuilder<CaseData, State, UserRole> configBuilder) {
@@ -72,8 +72,7 @@ public class CitizenInviteApplicant2 implements CCDConfig<CaseData, State, UserR
         data.getCaseInvite().setAccessCode(generateAccessCode());
         data.setDueDate(LocalDate.now().plus(2, ChronoUnit.WEEKS));
 
-        applicationSentForReviewApplicant1Notification.send(data, details.getId());
-        applicationSentForReviewApplicant2Notification.send(data, details.getId());
+        notificationDispatcher.send(applicationSentForReviewNotification, data, details.getId());
 
         return AboutToStartOrSubmitResponse.<CaseData, State>builder()
             .data(data)
