+++ src/main/java/uk/gov/hmcts/divorce/citizen/notification/ApplicationSentForReviewNotification.java
@@ -1,0 +1,79 @@
+package uk.gov.hmcts.divorce.citizen.notification;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+import uk.gov.hmcts.divorce.common.config.EmailTemplatesConfig;
+import uk.gov.hmcts.divorce.divorcecase.model.Applicant;
+import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
+import uk.gov.hmcts.divorce.notification.ApplicantNotification;
+import uk.gov.hmcts.divorce.notification.CommonContent;
+import uk.gov.hmcts.divorce.notification.NotificationService;
+
+import java.util.Map;
+
+import static uk.gov.hmcts.divorce.notification.CommonContent.ACCESS_CODE;
+import static uk.gov.hmcts.divorce.notification.CommonContent.CREATE_ACCOUNT_LINK;
+import static uk.gov.hmcts.divorce.notification.CommonContent.IS_REMINDER;
+import static uk.gov.hmcts.divorce.notification.CommonContent.NO;
+import static uk.gov.hmcts.divorce.notification.CommonContent.SUBMISSION_RESPONSE_DATE;
+import static uk.gov.hmcts.divorce.notification.CommonContent.YES;
+import static uk.gov.hmcts.divorce.notification.CommonContent.isDivorce;
+import static uk.gov.hmcts.divorce.notification.EmailTemplateName.JOINT_APPLICANT1_ANSWERS_SENT_FOR_REVIEW;
+import static uk.gov.hmcts.divorce.notification.EmailTemplateName.JOINT_APPLICANT2_ANSWERS_SENT_FOR_REVIEW;
+import static uk.gov.hmcts.divorce.notification.FormatUtil.DATE_TIME_FORMATTER;
+
+@Component
+@Slf4j
+public class ApplicationSentForReviewNotification implements ApplicantNotification {
+
+    public static final String APPLICANT_2_SIGN_IN_DIVORCE_URL = "applicant2SignInDivorceUrl";
+    public static final String APPLICANT_2_SIGN_IN_DISSOLUTION_URL = "applicant2SignInDissolutionUrl";
+
+    @Autowired
+    private NotificationService notificationService;
+
+    @Autowired
+    private CommonContent commonContent;
+
+    @Autowired
+    private EmailTemplatesConfig config;
+
+    @Override
+    public void sendToApplicant1(final CaseData caseData, final Long id) {
+        log.info("Sending application sent for review notification to applicant 1 for case : {}", id);
+
+        Map<String, String> templateVars =
+            commonContent.mainTemplateVars(caseData, id, caseData.getApplicant1(), caseData.getApplicant2());
+        templateVars.put(SUBMISSION_RESPONSE_DATE, caseData.getDueDate().format(DATE_TIME_FORMATTER));
+
+        notificationService.sendEmail(
+            caseData.getApplicant1().getEmail(),
+            JOINT_APPLICANT1_ANSWERS_SENT_FOR_REVIEW,
+            templateVars,
+            caseData.getApplicant1().getLanguagePreference()
+        );
+    }
+
+    @Override
+    public void sendToApplicant2(final CaseData caseData, final Long id) {
+        log.info("Sending application sent for review notification to applicant 2 for case : {}", id);
+
+        notificationService.sendEmail(
+            caseData.getApplicant2EmailAddress(),
+            JOINT_APPLICANT2_ANSWERS_SENT_FOR_REVIEW,
+            templateVars(caseData, id, caseData.getApplicant2(), caseData.getApplicant1(), false),
+            caseData.getApplicant1().getLanguagePreference()
+        );
+    }
+
+    private Map<String, String> templateVars(CaseData caseData, Long id, Applicant applicant, Applicant partner, boolean isReminder) {
+        Map<String, String> templateVars = commonContent.mainTemplateVars(caseData, id, applicant, partner);
+        templateVars.put(IS_REMINDER, isReminder ? YES :  NO);
+        templateVars.put(SUBMISSION_RESPONSE_DATE, caseData.getDueDate().format(DATE_TIME_FORMATTER));
+        templateVars.put(ACCESS_CODE, caseData.getCaseInvite().getAccessCode());
+        templateVars.put(CREATE_ACCOUNT_LINK,
+            config.getTemplateVars().get(isDivorce(caseData) ? APPLICANT_2_SIGN_IN_DIVORCE_URL : APPLICANT_2_SIGN_IN_DISSOLUTION_URL));
+        return templateVars;
+    }
+}
