--- src/test/java/uk/gov/hmcts/divorce/citizen/notification/ApplicationSentForReviewApplicant1NotificationTest.java
@@ -1,62 +1,0 @@
-package uk.gov.hmcts.divorce.citizen.notification;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
-import uk.gov.hmcts.divorce.divorcecase.model.Gender;
-import uk.gov.hmcts.divorce.notification.CommonContent;
-import uk.gov.hmcts.divorce.notification.NotificationService;
-
-import static org.hamcrest.Matchers.allOf;
-import static org.hamcrest.collection.IsMapContaining.hasEntry;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import static org.mockito.hamcrest.MockitoHamcrest.argThat;
-import static uk.gov.hmcts.divorce.divorcecase.model.LanguagePreference.ENGLISH;
-import static uk.gov.hmcts.divorce.notification.CommonContent.SUBMISSION_RESPONSE_DATE;
-import static uk.gov.hmcts.divorce.notification.EmailTemplateName.JOINT_APPLICANT1_ANSWERS_SENT_FOR_REVIEW;
-import static uk.gov.hmcts.divorce.notification.FormatUtil.DATE_TIME_FORMATTER;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_USER_EMAIL;
-import static uk.gov.hmcts.divorce.testutil.TestDataHelper.LOCAL_DATE;
-import static uk.gov.hmcts.divorce.testutil.TestDataHelper.caseData;
-import static uk.gov.hmcts.divorce.testutil.TestDataHelper.getApplicant2;
-import static uk.gov.hmcts.divorce.testutil.TestDataHelper.getMainTemplateVars;
-
-
-@ExtendWith(MockitoExtension.class)
-class ApplicationSentForReviewApplicant1NotificationTest {
-
-    @Mock
-    private NotificationService notificationService;
-
-    @Mock
-    private CommonContent commonContent;
-
-    @InjectMocks
-    private ApplicationSentForReviewApplicant1Notification notification;
-
-    @Test
-    void shouldSendEmailToApplicant1WhileInAwaitingApplicant2ResponseState() {
-        CaseData data = caseData();
-        data.setDueDate(LOCAL_DATE);
-        data.setApplicant2(getApplicant2(Gender.MALE));
-        when(commonContent.mainTemplateVars(data, 1234567890123456L, data.getApplicant1(), data.getApplicant2()))
-            .thenReturn(getMainTemplateVars());
-
-        notification.send(data, 1234567890123456L);
-
-        verify(notificationService).sendEmail(
-            eq(TEST_USER_EMAIL),
-            eq(JOINT_APPLICANT1_ANSWERS_SENT_FOR_REVIEW),
-            argThat(allOf(
-                hasEntry(SUBMISSION_RESPONSE_DATE, LOCAL_DATE.format(DATE_TIME_FORMATTER))
-            )),
-            eq(ENGLISH)
-        );
-        verify(commonContent).mainTemplateVars(data, 1234567890123456L, data.getApplicant1(), data.getApplicant2());
-    }
-}
